Added optional fast Sin/Cos trig lookup table support to entire codebase. Two playthroughs of entire game, one using and one without have verified there are no differences in gameplay between the two. Refined the trig lookup table mentioned above that had been added in a previous revision, a bit prematurely. Certain functions are no longer inlined. Added trig radians/degrees conversion macros to Math.h and added support to entire codebase. All Rad/Deg conversions now require only a single multiplication by a constant, instead of a multiply and divide. Moved PI #define from Globals.h to Math.h. While overhauling all trig math for above improvements, converted sqrt() and atan() calls to use the float versions sqrtf() & atanf() to avoid costly and unnecessary double<->float conversions. While overhauling all trig math for above improvements, converted double vars to floats when they weren't necessary, particularly in Partrikelsystem.cpp and Projectiles.cpp. Added #ifdef _DEBUG blocks around DebugMode sections. Disabled unused, empty function DrawSpecialLayer() in TileEngine.cpp/Gameplay.cpp. Fixed various uninitlaized variable warning in Projectiles.cpp.

+++ Hurrican/src/DX8Sound.cpp
@@ -965,7 +965,8 @@ bool CSoundManager::PlayWave3D(int x, int y, int Freq, int Nr)
     xdiff = ((pPlayer[0]->xpos + 45)  - x);
     ydiff = ((pPlayer[0]->ypos + 45)  - y);
 
-    Abstand = float(sqrt((xdiff * xdiff) + (ydiff * ydiff)));
+    //DKS - converted to float:
+    Abstand = sqrtf((xdiff * xdiff) + (ydiff * ydiff));
 
     vol = int(100-float(Abstand/6.0f));
     if (vol < 0)
@@ -1004,7 +1005,8 @@ void CSoundManager::Update3D (int x, int y, int Nr)
     xdiff = ((pPlayer[0]->xpos + 45)  - x);
     ydiff = ((pPlayer[0]->ypos + 45)  - y);
 
-    Abstand = float(sqrt((xdiff * xdiff) + (ydiff * ydiff)));
+    //DKS - converted to float:
+    Abstand = sqrtf((xdiff * xdiff) + (ydiff * ydiff));
 
     vol = (int)((100-float(Abstand/6.0f)) / 100.0f * pSoundManager->its_GlobalSoundVolume);
